pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "colordoll"
version = "0.1.8"
description = "Colorize text and data structures in the terminal with ANSI escape codes and themed decorators."
readme = {file = "README.md", content-type = "text/markdown"}
requires-python = ">=3.10"
license = "MIT"
authors = [
  {name = "kai gouthro"},
]
dependencies = ["pyyaml>=6.0.2"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/kaigouthro/colordoll"

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
where = ["src"]

